# docs/swagger.yaml
basePath: /
definitions:
...
  models.CategoryResponse:
    properties:
      id:
        example: 1
        type: integer
      image:
        example: phone.jpg
        type: string
      name:
        example: Điện thoại
        type: string
    type: object
  models.News:
    properties:
      author:
        type: string
      content:
        type: string
      created_at:
        type: string
      highlight:
        description: ✅ cần có
        type: string
      id:
        type: integer
      image:
        description: ảnh đại diện
        type: string
      tags:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 1
        type: integer
      items:
        description: ➕ Danh sách sản phẩm trong đơn hàng
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        example: Đang xử lý
        type: string
      total:
        example: 1500000
        type: number
      updated_at:
        type: string
      user_id:
        example: 2
        type: integer
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 1
        type: integer
      order_id:
        example: 10
        type: integer
      price:
        example: 750000
        type: number
      product_id:
        example: 5
        type: integer
      product_image:
        description: ➕ Trả riêng tên & ảnh sản phẩm
        example: laptop.jpg
        type: string
      product_name:
        example: Laptop Acer Gaming
        type: string
      quantity:
        example: 2
        type: integer
      updated_at:
        type: string
    type: object
...
  services.CreateOrderInput:
    properties:
      items:
        items:
          properties:
            product_id:
              type: integer
            quantity:
              type: integer
          type: object
        type: array
    type: object
  services.DashboardData:
    properties:
      top_products:
        items:
          $ref: '#/definitions/services.TopProductEntry'
        type: array
      total_orders:
        type: integer
      total_revenue:
        type: number
      total_users:
        type: integer
    type: object
  services.TopProductEntry:
    properties:
      name:
        type: string
      product_id:
        type: integer
      total_sold:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: you@example.com
    name: Developer
  description: API cho hệ thống bán hàng máy tính
  license:
    name: MIT
  termsOfService: http://localhost:8080/
  title: Computer Store API
  version: "1.0"
paths:
  /admin/dashboard:
    get:
      description: Trả về thống kê tổng quan hệ thống như tổng đơn hàng, doanh thu,
        số lượng người dùng, v.v.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DashboardData'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lấy dữ liệu dashboard admin
      tags:
      - Admin
...
  /categories:
    get:
      description: Trả về danh sách tất cả danh mục hiện có
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lấy tất cả danh mục
      tags:
      - Categories
    post:
      consumes:
      - multipart/form-data
      description: Tạo mới danh mục với tên và ảnh đại diện
      parameters:
      - description: Tên danh mục
        in: formData
        name: name
        required: true
        type: string
      - description: Ảnh đại diện danh mục
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tạo danh mục mới
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Xoá danh mục theo ID
      parameters:
      - description: ID danh mục
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Xoá danh mục
      tags:
      - Categories
    put:
      consumes:
      - multipart/form-data
      description: Cập nhật thông tin danh mục theo ID, có thể đổi tên và/hoặc ảnh
      parameters:
      - description: ID danh mục
        in: path
        name: id
        required: true
        type: string
      - description: Tên mới
        in: formData
        name: name
        type: string
      - description: Ảnh mới (tuỳ chọn)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật danh mục
      tags:
      - Categories
...
  /news:
    get:
      description: Trả về danh sách toàn bộ tin tức
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.News'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lấy danh sách tất cả tin tức
      tags:
      - News
    post:
      consumes:
      - multipart/form-data
      description: Tạo một tin tức mới từ form-data (bao gồm ảnh, tiêu đề, nội dung...)
      parameters:
      - description: Tiêu đề
        in: formData
        name: title
        required: true
        type: string
      - description: Nội dung
        in: formData
        name: content
        required: true
        type: string
      - description: Nội dung nổi bật
        in: formData
        name: highlight
        type: string
      - description: Tags (ngăn cách bằng dấu phẩy)
        in: formData
        name: tags
        type: string
      - description: Ảnh đại diện
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.News'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tạo tin tức mới
      tags:
      - News
  /news/{id}:
    delete:
      description: Xóa tin tức theo ID
      parameters:
      - description: ID tin tức
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Xóa tin tức
      tags:
      - News
    get:
      description: Trả về chi tiết tin tức theo ID
      parameters:
      - description: ID tin tức
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lấy chi tiết tin tức theo ID
      tags:
      - News
    put:
      consumes:
      - multipart/form-data
      description: Cập nhật tin tức theo ID (form-data)
      parameters:
      - description: ID tin tức
        in: path
        name: id
        required: true
        type: integer
      - description: Tiêu đề
        in: formData
        name: title
        type: string
      - description: Nội dung
        in: formData
        name: content
        type: string
      - description: Nội dung nổi bật
        in: formData
        name: highlight
        type: string
      - description: Tags
        in: formData
        name: tags
        type: string
      - description: Ảnh mới
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật tin tức
      tags:
      - News
  /orders:
    get:
      description: Trả về toàn bộ đơn hàng trong hệ thống (chỉ dành cho admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lấy tất cả đơn hàng
      tags:
      - Orders
  /orders/from-cart:
    post:
      description: Tạo đơn hàng dựa trên các sản phẩm trong giỏ hàng
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tạo đơn hàng từ giỏ hàng
      tags:
      - Orders
  /orders/manual:
    post:
      consumes:
      - application/json
      description: Người dùng tạo đơn hàng với thông tin tùy chỉnh (không qua giỏ
        hàng)
      parameters:
      - description: Thông tin đơn hàng
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.CreateOrderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tạo đơn hàng thủ công
      tags:
      - Orders
  /orders/user:
    get:
      description: Trả về danh sách đơn hàng của người dùng đã đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lấy đơn hàng của người dùng hiện tại
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"